backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
PyMuPDF>=1.23.0
msgspec>=0.13.0
spacy>=3.6.0
jinja2>=3.1.0
mistralai>=0.4.0
python-dotenv>=1.0.0
ruff>=0.1.0
